feat(payment): maps the payment fields and hbox
link payment methods to TextFields and to Hbox elements

adds event to TextField payment elements which create a new Payment object and set payment value.

adds event to the TextField payment for call the process payment method

also adds event to Hbox Elements which representes payment methods that create a new payment and set payment value

feat(payment): adds requestPayment method to handle payment input field and associate them with payment method

This requestPayment method ensures the paymentField is properly initialized and formatted. If the field is blank, it sets a default value of "0,00" and updates the payment method accordingly.

feat(payment): add processPayment and setPayment methods for handling and validating payment processes

The processPayment method ensure that the sale is valid by checking for null values and verifying the total price.  It then retrieves the unique payments from the payment methods and sets them to the sale. The method calcultes the total amount of payments and compares it with the sale's total price, throwing an exception of the total paument amount is less.

The setPaymentValue method updates the payment value for a given payment methods, It formats the input value, converts it to BigDecimal . and updates the payment method in the paymentMethods map. If the input value cannot be parsed, a runtime exception is thrown

feat(utils): add formatCurrency method for parsing currency strings

The formatCurrency method converts a string representtion of a currency value into a BigDecimal