refactor(service): replace enum with dedicated class in processSale
Replaced the direct usage of the `TypeCheckoutMovement` enum with a dedicated `CheckoutMovementType` class in the `processSale` method. This change improves code extensibility by encapsulating the movement type into a class, allowing for future additions of logic related to movement types.

BREAKING CHANGE: `checkoutMovementService.buildCheckoutMovement` now expects a `CheckoutMovementType` object instead of an enum.